move id generation to propagation
