+++ lib/api/libhoney.js
@@ -3,10 +3,10 @@ const libhoney = require("libhoney"),
   os = require("os"),
   process = require("process"),
   path = require("path"),
-  crypto = require("crypto"),
   url = require("url"),
   tracker = require("../async_tracker"),
   instrumentation = require("../instrumentation"),
+  { generateSpanId, generateTraceId } = require("../propagation"),
   deterministicSampler = require("../deterministic_sampler"),
   schema = require("../schema"),
   Span = require("./span"),
@@ -18,21 +18,6 @@ const defaultName = "nodejs";
 const incr = (payload, key, val = 1) => {
   payload[key] = (payload[key] || 0) + val;
 };
-
-// id generation
-const SPAN_ID_BYTES = 8;
-const TRACE_ID_BYTES = 16;
-
-function generateTraceId() {
-  crypto.randomBytes(TRACE_ID_BYTES).toString("hex");
-}
-
-function generateSpanId() {
-  crypto.randomBytes(SPAN_ID_BYTES).toString("hex");
-}
-
-exports.generateSpanId = generateSpanId;
-exports.generateTraceId = generateTraceId;
 
 const getFilteredOptions = opts => {
   let filteredOptions = {};
@@ -106,7 +91,7 @@ module.exports = class LibhoneyEventAPI {
   }
 
   startTrace(metadataContext, traceId, parentSpanId, dataset) {
-    const id = traceId || instrumentation.generate;
+    const id = traceId || generateTraceId();
 
     tracker.setTracked({
       id,
